Fall Goals Report p 1

Fall Goals Report p 1

Usage

fall_goals_report_p1(
  mapvizieR_obj,
  studentids,
  measurementscale,
  context,
  small_n_cutoff = 0,
  start_fws = "Spring",
  start_year_offset = -1,
  end_fws = "Spring",
  end_academic_year = 2015,
  entry_grade_seasons = c(-0.8, 4.2),
  exclude_prior_year_holdover = TRUE,
  detail_academic_year = 9999,
  goal_cgp = 80,
  school_type = "MS",
  localization = localize("Newark")
)

Arguments

mapvizieR_obj

a valid mapvizieR_object

studentids

a vector of studentids to run

measurementscale

desired subject

context

what school/grade/class/etc grouping is represented?

small_n_cutoff

drop a grade_level_season if less than x (useful when dealing with weird cohort histories)

start_fws

character, starting season for school growth norms

start_year_offset

0 if start season is same, -1 if start is prior year.

end_fws

ending season

end_academic_year

ending academic year

entry_grade_seasons

for becca plot. default is c(-0.8, 4.2)

exclude_prior_year_holdover

if, for instance, this year's 3rd grade includes students who were in the 3rd grade last year, drop their records from the data

detail_academic_year

passed through to various plots

goal_cgp

what target CGP should be used for goals?

school_type

c('ES', 'MS')

localization

controls goals and goal logic for report. See localization.R for more details

Value

a multipage report, represented as a list of grobs.