quealy_subgroups

kevin quealy of the nytimes did a nice job capturing change in the general population vs change in specific subgroups: http://nyti.ms/1tQrOIl re: obamacare. let's do the same thing for change in RIT score.

Usage

quealy_subgroups(
  mapvizieR_obj,
  studentids,
  measurementscale,
  subgroup_cols = c("starting_quartile"),
  pretty_names = c("Starting Quartile"),
  magic_subgroups = TRUE,
  start_fws,
  start_year_offset,
  end_fws,
  end_academic_year,
  start_fws_prefer = NA,
  report_title = NA,
  complete_obsv = TRUE,
  drop_NA_groups = TRUE,
  include_all = TRUE,
  small_n_cutoff = -1,
  join_by_measurementscale = FALSE,
  school_growth_norms = 2015
)

Arguments

mapvizieR_obj

mapvizieR object

studentids

target students

measurementscale

target subject

subgroup_cols

what subgroups in mapvizier roster do you want to cut by? default is starting_quartile

pretty_names

nicely formatted names for the column cuts used above.

magic_subgroups

generated by mapvizieR. possible values include: 'starting_quartile'

start_fws

one academic season (if known); pass vector of two and quealy subgroups will pick

start_year_offset

0 if start season is same, -1 if start is prior year.

end_fws

ending season

end_academic_year

ending academic year

start_fws_prefer

which term is preferred? not required if only one start_fws is passed

report_title

text grob to put on the report tile

complete_obsv

if TRUE, limit only to students who have BOTH a start and end score. default is TRUE.

drop_NA_groups

should we ignore subgroups with value NA? default is true

include_all

should the output have plot at the top showing the TOTAL variation? not recommended for data spanning multiple grade levels.

small_n_cutoff

drop a subgroup if less than x (useful for cutting off the long tail of a group). applies to all subgroups in subgroup_cols. does not apply to magic subgroups.

join_by_measurementscale

boolean, passed to roster_to_growth_df. TRUE if the subgroup col is per-subject.

school_growth_norms

passed through to calc_cgp. determines what norm study to use. default is 2015 (most recent study).

Value

a grob composed of multiple ggplots